style(vst3-host): clear the clippy -D warnings backlog
Ten lints, all pre-existing and mechanical: two redundant casts, a doc
list run into the paragraph below it, a test module carrying both an
outer and an inner doc comment, an impl sitting after `mod tests`, an
`EventList::is_empty` with no callers, and a loop index in the audio
oracle. No behaviour changes; suites unchanged at 135 unit + 22
conformance + 9 audio + 4 dispatch + 5 teardown.

`cargo clippy --all-targets --features conformance -- -D warnings` now
exits clean, so the gate is usable in CI instead of always-red.

+++ crates/plugin/formats/tutti-vst3-host/src/com/run_loop.rs
@@ -219,7 +219,7 @@ impl RunLoop {
         let mut state = self.state.lock().unwrap_or_else(|p| p.into_inner());
         // Zero would mean "every iteration"; clamp so a misbehaving plugin
         // cannot spin the host's UI loop.
-        let period = Duration::from_millis(milliseconds.max(1) as u64);
+        let period = Duration::from_millis(milliseconds.max(1));
         state.timers.insert(
             handler as usize,
             TimerEntry {
